Use float instead of double for the hot geometry math
Vector components and the intersection/lighting/shading math run once per
pixel (times reflections and shadow rays), so halving their width cuts
memory traffic and lets the vectoriser pack twice as many lanes. Switch
Vector, Triangle reflectivity/intersection-t, and the per-pixel locals in
main.cpp to float (FLT_MAX/FLT_EPSILON/sqrtf). Camera angles stay double
since they accumulate across frames.

+++ src/triangle.cpp
@@ -9,7 +9,7 @@ Triangle::Triangle()
 	calcInternalVectors();
 }
 
-Triangle::Triangle(const Vector& _v0, const Vector& _v1, const Vector& _v2, Colour _colour, bool _reflective, double _reflectivity, bool _invisible, bool _glows)
+Triangle::Triangle(const Vector& _v0, const Vector& _v1, const Vector& _v2, Colour _colour, bool _reflective, float _reflectivity, bool _invisible, bool _glows)
 :v0(_v0), v1(_v1), v2(_v2), colour(_colour), reflective(_reflective), reflectivity(_reflectivity), invisible(_invisible), glows(_glows)
 {
 	calcInternalVectors();
@@ -30,23 +30,23 @@ void Triangle::calcInternalVectors()
 	normal.normalise();
 }
 
-bool Triangle::getIntersection(const Ray& ray, Vector& ret, double& t_ret, Colour& colour_ret)
+bool Triangle::getIntersection(const Ray& ray, Vector& ret, float& t_ret, Colour& colour_ret)
 {
 	Vector h = ray.direction.crossprod(v0v2);
-	double a = v0v1.dotprod(h);
-	if (fabs(a) < DBL_EPSILON) {return false;}
-	
-	double f =  1/a;
+	float a = v0v1.dotprod(h);
+	if (fabs(a) < FLT_EPSILON) {return false;}
+
+	float f =  1/a;
 	Vector s = ray.start - v0;
-	double u = f * s.dotprod(h);
+	float u = f * s.dotprod(h);
 	if (u < 0 || u > 1) {return false;}
-	
+
 	Vector q = s.crossprod(v0v1);
-	double v = f * ray.direction.dotprod(q);
+	float v = f * ray.direction.dotprod(q);
 	if (v < 0 || u + v > 1) {return false;}
-	
-	double t = f * v0v2.dotprod(q);
-	if (t > DBL_EPSILON)
+
+	float t = f * v0v2.dotprod(q);
+	if (t > FLT_EPSILON)
 	{
 		ret = ray.start + ray.direction * t;
 		t_ret = t;
