+++ src/vector.cpp
@@ -6,7 +6,7 @@ Vector::Vector()
 :x(0), y(0), z(0)
 {}
 
-Vector::Vector(double _x, double _y, double _z)
+Vector::Vector(float _x, float _y, float _z)
 :x(_x), y(_y), z(_z)
 {}
 
@@ -14,9 +14,9 @@ Vector::Vector(const Vector& rhs)
 :x(rhs.x), y(rhs.y), z(rhs.z)
 {}
 
-double Vector::length() const {return std::sqrt(x*x + y*y + z*z);}
-double Vector::lengthsq() const {return x*x + y*y + z*z;}
-double Vector::dotprod(const Vector& rhs) const
+float Vector::length() const {return std::sqrt(x*x + y*y + z*z);}
+float Vector::lengthsq() const {return x*x + y*y + z*z;}
+float Vector::dotprod(const Vector& rhs) const
 {
 	return x*rhs.x + y*rhs.y + z*rhs.z;
 }
@@ -28,7 +28,7 @@ Vector Vector::crossprod(const Vector& rhs) const
 
 void Vector::normalise()
 {
-	double l = length();
+	float l = length();
 	if (l != 0) {*this /= l;}
 }
 
@@ -54,23 +54,23 @@ Vector& Vector::operator-=(const Vector& rhs)
 	return *this;
 }
 
-Vector Vector::operator*(double f) const
+Vector Vector::operator*(float f) const
 {
 	return Vector(x*f, y*f, z*f);
 }
 
-Vector Vector::operator/(double f) const
+Vector Vector::operator/(float f) const
 {
 	return Vector(x/f, y/f, z/f);
 }
 
-Vector& Vector::operator*=(double f)
+Vector& Vector::operator*=(float f)
 {
 	*this = *this * f;
 	return *this;
 }
 
-Vector& Vector::operator/=(double f)
+Vector& Vector::operator/=(float f)
 {
 	*this = *this / f;
 	return *this;
@@ -89,15 +89,15 @@ bool Vector::operator!=(const Vector& rhs) const
 void Vector::rotateAroundX(double rad)
 {
 	double y_t = y;
-	
-	y = y * cos(rad) - z * sin(rad);
-	z = y_t * sin(rad) + z * cos(rad);
+
+	y = static_cast<float>(y * cos(rad) - z * sin(rad));
+	z = static_cast<float>(y_t * sin(rad) + z * cos(rad));
 }
 
 void Vector::rotateAroundY(double rad)
 {
 	double x_t = x;
-	
-	x = x * cos(rad) - z * sin(rad);
-	z = x_t * sin(rad) + z * cos(rad);
+
+	x = static_cast<float>(x * cos(rad) - z * sin(rad));
+	z = static_cast<float>(x_t * sin(rad) + z * cos(rad));
 }
